node-runner-cli/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
-e git+https://github.com/radixdlt/python-core-client@1a61f8a1e2acae859522bbff923b0042ec8955e5#egg=core-client
bech32==1.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
cryptography==36.0.1
deepdiff==5.8.0
deepmerge==0.3.0
ecdsa==0.17.0
idna==3.3; python_version >= '3'
ordered-set==4.1.0; python_version >= '3.7'
pycparser==2.21
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.0
requests==2.26.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
